setup hibernate & add loadDB func
