+++ build.gradle
@@ -1,6 +1,7 @@
 plugins {
   id 'java'
   id 'war'
+  id("io.freefair.lombok") version "8.4"
 }
 
 group 'aca98b'
