+++ src/main/java/aca98b/web3lv2/beans/OneElement.java
@@ -1,0 +1,80 @@
+package aca98b.web3lv2.beans;
+
+public final class OneElement {
+    public float x;
+    public float y;
+    public float r;
+    public String result;
+    public String time;
+    public String scriptTime;
+    public String uid;
+
+    public OneElement(float x, float y, float r, String result, String time, String scriptTime, String uid){
+        this.x = x;
+        this.y = y;
+        this.r = r;
+        this.result = result;
+        this.time = time;
+        this.scriptTime = scriptTime;
+        this.uid = uid;
+    }
+
+    public OneElement(){
+    }
+//
+//    public float getX() {
+//        return x;
+//    }
+//
+//    public void setX(float x) {
+//        this.x = x;
+//    }
+//
+//    public float getY() {
+//        return y;
+//    }
+//
+//    public void setY(float y) {
+//        this.y = y;
+//    }
+//
+//    public float getR() {
+//        return r;
+//    }
+//
+//    public void setR(float r) {
+//        this.r = r;
+//    }
+//
+//    public String getResult() {
+//        return result;
+//    }
+//
+//    public void setResult(String result) {
+//        this.result = result;
+//    }
+//
+//    public String getTime() {
+//        return time;
+//    }
+//
+//    public void setTime(String time) {
+//        this.time = time;
+//    }
+//
+//    public String getScriptTime() {
+//        return scriptTime;
+//    }
+//
+//    public void setScriptTime(String scriptTime) {
+//        this.scriptTime = scriptTime;
+//    }
+//public String getUid() {
+//        return uid;
+//    }
+//
+//    public void setUid (String uid) {
+//        this.uid = uid;
+//    }
+
+}
