+++ src/main/webapp/views/mainPage.xhtml
@@ -77,7 +77,7 @@
                             <h:inputText id="x" value="#{xBean.value}" maxlength="7"
                                          validator="#{xBean.xChecker}">
                                 <f:ajax execute="x" render=":msgPanel"/>
-<!--                                <f:ajax execute="x" render="@form"/>-->
+                                <!--                                <f:ajax execute="x" render="@form"/>-->
                             </h:inputText>
                         </div>
                         <div class="y-input" id="y-input">
